refactor: move FormField to inputs
