+++ components/forms/index.ts
@@ -1,2 +1,1 @@
 export { default as BaseForm } from './BaseForm';
-export { default as FormField } from './FormField';
