+++ components/inputs/index.ts
@@ -1,0 +1,1 @@
+export { default as FormField } from './FormField';
